dashboard/requirements.txt
streamlit>=1.35.0
plotly>=5.22.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
spotipy>=2.24.0